refactor(search): Split DuckDB vector index module into separate files
Break up the monolithic duckdb.rs into focused submodules:
- metric.rs: DuckDBDistanceMetric enum and distance/score expressions
- hnsw.rs: DuckDBHnswOptions and HNSW index SQL generation
- sql.rs: Vector search SQL builders (CTE and flat paths), filter pushdown
- query_exec.rs: DuckDBVectorQueryExec ExecutionPlan implementation
- query_table.rs: DuckDBVectorQueryTable TableProvider implementation
- mod.rs: DuckDBVectorIndex, trait impls, shared utilities, re-exports

Also improves DuckDBVectorQueryExec display to show metric, dims,
scan type (hnsw vs brute_force), and projection columns.
